requirement_all_common.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
branca==0.3.0
cached-property==1.4.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.2.4
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-crispy-forms-registration==0.1.3
django-oauth-toolkit==1.2.0
django-oidc-provider==0.7.0
django-python3-ldap==0.11.2
django-registration==3.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
docker==3.5.0
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
folium==0.6.0
future==0.16.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Jinja2==2.10
jsonschema==2.6.0
ldap==1.0.2
ldap3==2.5
lxml==4.2.4
MarkupSafe==1.0
numpy==1.15.2
oauthlib==2.1.0
parsel==1.5.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycryptodomex==3.6.6
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyjwkest==1.4.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
pypiwin32==223
pytils==0.3
pytz==2018.5
pywin32==223
PyYAML==3.13
queuelib==1.5.0
requests==2.18.4
reverse==0.1.0
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
texttable==0.9.1
tools==0.1.9
Twisted==18.7.0
urllib3==1.22
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
w3lib==1.19.0
websocket-client==0.51.0
zope.interface==4.5.0
